requirements.txt
asgiref==3.8.1
Django==4.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
